% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Constant returns to scale production function}
\usage{
cd(K, L, theta = 2/3, A = 1)
}
\arguments{
\item{K}{capital}

\item{L}{labor}

\item{A}{total factor}
}
\description{
Constant returns to scale production function
}
\examples{
cd(10, 100)
}
\author{
GONG Kuiyuan
}
\keyword{Production}
\keyword{function}
