% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_plot.R
\name{conv_plot}
\alias{conv_plot}
\title{Plot for convergence}
\usage{
conv_plot(x, m)
}
\arguments{
\item{x}{\code{numeric vector} raw estimates of parameters}

\item{m}{\code{numeric} number of iterations}
}
\value{
plot
}
\description{
Prepare a histogram ggplot for posterior distribution of a given parameter. To be used within codes for plotting grids of posterior distributions of several parameters.
}
\examples{
conv_plot(rnorm(1500), m =500)
}
