% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_analysis.R
\name{Gro_analysis}
\alias{Gro_analysis}
\title{Growth model selection}
\usage{
Gro_analysis(
  data_weight,
  all_mods = c("vonbertalanffy"),
  random = list(),
  logtransform = FALSE,
  run = list(nit = 100, nburnin = 10, nthin = 1, nch = 1),
  parallel = FALSE
)
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the numeric columns \emph{Age}, \emph{MeasurementValue} and \emph{AnimalAnonID}}

\item{all_mods}{\code{vector of character} indicating the growth models that need to be fit.The following models are supported :logistic, gompertz, tpgm, power, richards, vonbertalanffy. default = "vonBertalanffy"}

\item{random}{{list} of the model names giving the parameters that should include an individual random effect. See the example}

\item{logtransform}{{logical} whether age and measurement values should be log transform (+1 is added to avoid having log(0))}

\item{run}{\code{list} Bayesian parameters. They should be increased to reach convergence \itemize{
\item \code{nch} number of chains.
\item \code{nthin} interval between iterations to keep.
\item \code{nburnin} number of iterations to discard.
\item \code{nit} total number of iterations.}}

\item{parallel}{\code{logical} Whether the model should be run in parallel}
}
\value{
a list including:
\itemize{
\item model: the fit of the best model
\item the wAIC table of the model
}
}
\description{
This function fit a series of growth models to a dataset and select the best one by wAIC.
}
\examples{
Age <- sample(c(0:10), 1000, replace = TRUE)
MeasurementValue <- exp(0.2+15 * (1 - exp(-(0.1) * log(Age+1)))+ rnorm(1000,0,0.01))-1 
AnimalAnonID <- sample(c(0:20), 100, replace = TRUE)
dat = data.frame(Age = Age, MeasurementValue = MeasurementValue, 
                 AnimalAnonID = AnimalAnonID, MeasurementType = "Live Weight")

#Test 4 models: vonbertalanffy including an individual random effect on z0
#               vonbertalanffy including individual random effects on z0 and zinf
#               gompertz including an individual random effect on gamma 
#               gompertz including no individual random effect
a = Gro_analysis(dat, all_mods  = c("vonbertalanffy", "gompertz"),
                 random = list(vonbertalanffy = c("z0", "z0, zinf"), gompertz = c("gamma", "")),
                 run = list(nit = 1000, nburnin = 100, nthin = 1, nch = 1))
}
