% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_pred.R
\name{Gro_pred}
\alias{Gro_pred}
\title{Predictions and plots}
\usage{
Gro_pred(data, out, Cred_int = c(0.025, 0.975), title = "")
}
\arguments{
\item{data}{\code{data.frame} including at least the numeric columns \emph{Age}, \emph{MeasurementValue} and \emph{AnimalAnonID}}

\item{out}{result output from Gro_run()}

\item{Cred_int}{\code{numeric} lower and upper bound of the credible interval to plot and to predict}

\item{title}{\code{character} Title of the plot}
}
\value{
a list including \itemize{
\item \code{summary}shomwing mean, standard deviation, credible interval at 95\% and the Gelman-Rubin statistics Rhat of each parameter.
\item \code{predictions} giving the mean estimates and credible interval at 95\% of each age
\item \code{GOF}: 4 tests of googness of fit: "normal": test if residuals are normally distributes ; "X"; test if there is a trend between residual and Age, "var": Test if there is the trend in the variance of the residuals over Age ,"conv": check if all Rhat < 1.1
\item \code{plot_pred} Plot of the predicted values, credible interval at 95\% in grey and data points.
\item \code{convergence}: Plots of the Bayesian chains
\item \code{posterior}: Plots of the posterior distribution for each parameter
}
}
\description{
Give predicted values of the model and plots to check the convergence of the model
}
\examples{
Age <- sample(c(0:10), 100, replace = TRUE)
AnimalAnonID <- sample(c(0:20), 100, replace = TRUE)
MeasurementValue <- exp(0.2+15 * (1 - exp(-(0.1) * log(Age+1)))+ 
                          rnorm(100,0,0.01) + AnimalAnonID*0.1)-1 
dat = data.frame(Age = Age, MeasurementValue = MeasurementValue, 
                 AnimalAnonID = AnimalAnonID, MeasurementType = "Live Weight")

out = Gro_analysis(dat, 
                   all_mods = c("vonbertalanffy"),
                   run = list(nit = 1000, nburnin = 100, nthin = 1, nch = 3))

p <- Gro_pred(data = dat, 
              out = out$model, 
              title =out$wAIC_tab$model[1])
p$summary
p$predictions
p$GOF
p$plot_pred
p$convergence
p$posterior
}
