% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_run.R
\name{Gro_run}
\alias{Gro_run}
\title{Run growth model}
\usage{
Gro_run(
  x,
  dat,
  all_mods,
  random = "",
  run = list(nit = 100, nburnin = 10, nthin = 1, nch = 1)
)
}
\arguments{
\item{x}{\code{numeric} index of the formula used to build the model}

\item{dat}{\code{data.frame}including at least the numeric columns \emph{age}, \emph{z} and \emph{IND}}

\item{all_mods}{\code{vector of characters} of model names. The following models are supported: logistic, gompertz, tpgm, power, richards, vonbertalanffy.}

\item{random}{{vector of character} of the same length as \code{all_mods} giving the parameters that should be included an individual random effect}

\item{run}{\code{list} Bayesian parameters. They should be increased to reach convergence \itemize{
\item \code{nch} number of chains.
\item \code{nthin} interval between iterations to keep.
\item \code{nburnin} number of iterations to discard.
\item \code{nit} total number of iterations.
}}
}
\value{
This function returns a \code{list}:
\itemize{
\item \code{model} a list including estimates of coefficients and model characteristics
\item \code{tab} a data frame with information on models & WAIC
}
}
\description{
Build and run the bayesian growth model \code{x} of \code{all_mods}.
}
\examples{
age <- rnorm(10000, 0, 1)
id1 =  rnorm(21,0, 0.5)
id2 =  rnorm(21,0, 0.4)
id3 =  rnorm(21,0, 0.3)
IND =sample(c(1:20), 100, replace = TRUE)
z <- 0.2+ id1[IND]+ (15 + id2[IND])* (1 - exp(-(1+ id3[IND]) * age)) +
  rnorm(100, 0, 0.01)
dat = data.frame(age = age, z = z, 
                 IND = as.numeric(factor(IND ,labels = c(1:length(unique(IND)))))
)

#Run a vonbertalanffy model including an individual effect on z0
out = Gro_run(1, 
              dat,
              all_mods  = c("vonbertalanffy"),
              random = c("z0"),
              run = list(nit = 500, nburnin = 100, nthin = 10, nch = 3))
out$tab

}
