% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_writenimblecode.R
\name{Gro_writenimblecode}
\alias{Gro_writenimblecode}
\title{Write nimble code for growth model}
\usage{
Gro_writenimblecode(
  params,
  model,
  random = c(),
  maxval = list(),
  minval = list()
)
}
\arguments{
\item{params}{\code{vector of character} names of all parameters of the model.}

\item{model}{\code{character} likelihood of the nimble model}

\item{random}{\code{vector of character} name of the parameters that must include an individual random effect}

\item{maxval}{\code{list} including the maximum values for prior for each parameter. Default value is 1000.}

\item{minval}{\code{list} including the minimum values for prior for each parameter.Default value is 0.}
}
\value{
A character representing a nimble model
}
\description{
Write the nimble code needed to run growth model including random effects
}
\examples{
#Example for a von bertalanffy model
Gro_writenimblecode(params = c('z0', 'zinf', 'gamma'), 
                    model = "for (j in 1:N){{
       logz[j] ~dnorm(z0[IND[j]] + zinf[IND[j]] * 
       (1 - exp(- gamma[IND[j]] * logx[j])), sigma_res)}}",
                    random = "gamma",
                    maxval = list(z0 = 5), 
                    minval= list(z0 = 0, zinf = 0, gamma = 0)
)
}
