% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth-package.R
\docType{package}
\name{Growth-package}
\alias{Growth}
\alias{Growth-package}
\title{Growth: Bayesian Growth Analysis}
\description{
Use bayesian approach to builg growth model with random effect.
}
\author{
\strong{Maintainer}: Floriane Plard \email{floriane.c.plard@gmail.com} [contractor]

Other contributors:
\itemize{
  \item Species 360 \email{Support@Species360.org} [copyright holder]
}

}
\keyword{internal}
