\name{growth}
\title{Normal Theory Repeated Measurements Library}
\description{
\code{\link{carma}}			Normal Theory Repeated Measurements Models

\code{\link{corgram}}			Calculate and Plot a Correlogram

\code{\link{elliptic}}		Multivariate Elliptically-contoured Nonlinear Repeated Measurements Models with Time-varying Covariates

\code{\link{pergram}}			Calculate and Plot a Periodogram

\code{\link{potthoff}}		Potthoff and Roy Growth Curve Model

\code{\link{rmaov}}			Split-plot ANOVA Model
}
