% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_post.R
\name{hist_post}
\alias{hist_post}
\title{Histogram}
\usage{
hist_post(x, namex = "", namelab = "")
}
\arguments{
\item{x}{\code{numeric vector} raw estimates of parameters}

\item{namex}{\code{character} x axis label}

\item{namelab}{\code{character} title label of the plot}
}
\value{
histogram plot
}
\description{
Prepare a histogram ggplot for posterior distribution of a given parameter. To be used within codes for plotting grids of posterior distributions of several parameters.
}
\examples{
hist_post(rnorm(1000))
}
