% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhatfun.R
\name{Rhatfun}
\alias{Rhatfun}
\title{Estimate Rhat values}
\usage{
Rhatfun(rb, nch, it, nparam)
}
\arguments{
\item{rb}{\code{array of dimension 3} estimates of output chains of a bayesian model. The rows should be the iterations, the column the different chains and the 3rd dimension the parameters}

\item{nch}{\code{numeric} number of chains}

\item{it}{\code{numeric} number of iterations}

\item{nparam}{\code{numeric} number of parameters}
}
\value{
\code{numeric vector} of size \code{nparam} giving Rhat statistics for each parameter
}
\description{
Derived Rubin Gelman convergence statistics from output chains of a Bayesian model
}
\examples{
rb = array(rnorm(15*3*4), dim = c(15,3,4))
Rhatfun(rb, nch = 3, it = 15, nparam = 4)

}
