% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhatfun.R
\name{sum_nim}
\alias{sum_nim}
\title{Summary statistics}
\usage{
sum_nim(rb2, nch)
}
\arguments{
\item{rb2}{\code{array} of dimension 3 or 2 including estimates of output chains of a bayesian model. If of dimension 3, the rows should be the iterations, the columns the different chains and the 3rd dimension the parameters. If of dimension 2, the rows should be the iterations and chains, and the columns the parameters.}

\item{nch}{\code{numeric} number of chains}
}
\value{
A summary matrix with columns giving mean, sd, credible interval at 2.5\% and 97.5\%, and Rubin Gelman Rhat statistic. The different parameters are on lines.
}
\description{
Makes a summary table with main statistics for each parameter from output chains of a Bayesian model
}
\examples{
rb = array(rnorm(15*3*4), dim = c(15,3,4))
sum_nim(rb, nch = 3)

}
