\name{growth}
\title{Normal Theory Repeated Measurements Library}
\alias{growth}
\description{
\code{\link[growth]{carma}}			Normal Theory Repeated Measurements Models

\code{\link[growth]{corgram}}			Calculate and Plot a Correlogram

\code{\link[growth]{elliptic}}		Multivariate Elliptically-contoured Nonlinear Repeated Measurements Models with Time-varying Covariates

\code{\link[growth]{pergram}}			Calculate and Plot a Periodogram

\code{\link[growth]{potthoff}}		Potthoff and Roy Growth Curve Model

\code{\link[growth]{rmaov}}			Split-plot ANOVA Model
}
\keyword{documentation}
