% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor.r
\name{pergram}
\alias{pergram}
\alias{plot.cum.pergram}
\alias{plot.pergram}
\title{Calculate and Plot a Periodogram}
\usage{
pergram(y)
}
\arguments{
\item{y}{A time series vector.}

\item{add}{If TRUE, adds a new periodogram to an existing plot.}

\item{others}{Plotting parameters}
}
\value{
\code{pergram} prints and returns a two-column matrix of class,
\code{pergram}, containing the periodogram.
}
\description{
\code{pergram} calculates the values of a periodogram, \code{plot.pergram}
plots it, and \code{plot.cum.pergram} plots the corresponding cumulative
periodogram.
}
\examples{

y <- rnorm(100)
print(z <- pergram(y))
plot(z)
plot.cum(z)

}
\author{
J.K. Lindsey
}
\keyword{hplot}

